Add macOS Application
