+++ .gitignore
@@ -1,5 +1,6 @@
 .DS_Store
-/.build
-/Packages
-/*.xcodeproj
+.build
+Packages
+*.xcodeproj
 .swiftpm
+xcuserdata
