+++ OrgChart Generator/AppDelegate.swift
@@ -1,0 +1,39 @@
+//
+//  AppDelegate.swift
+//  OrgChart Generator
+//
+//  Created by Paul Schmiedmayer on 5/19/20.
+//  Copyright © 2020 Paul Schmiedmayer. All rights reserved.
+//
+
+import Cocoa
+import SwiftUI
+
+@NSApplicationMain
+class AppDelegate: NSObject, NSApplicationDelegate {
+
+    var window: NSWindow!
+
+
+    func applicationDidFinishLaunching(_ aNotification: Notification) {
+        // Create the SwiftUI view that provides the window contents.
+        let contentView = ContentView()
+
+        // Create the window and set the content view. 
+        window = NSWindow(
+            contentRect: NSRect(x: 0, y: 0, width: 480, height: 300),
+            styleMask: [.titled, .closable, .miniaturizable, .resizable, .fullSizeContentView],
+            backing: .buffered, defer: false)
+        window.center()
+        window.setFrameAutosaveName("Main Window")
+        window.contentView = NSHostingView(rootView: contentView)
+        window.makeKeyAndOrderFront(nil)
+    }
+
+    func applicationWillTerminate(_ aNotification: Notification) {
+        // Insert code here to tear down your application
+    }
+
+
+}
+
