+++ OrgChart Generator/ContentView.swift
@@ -1,0 +1,23 @@
+//
+//  ContentView.swift
+//  OrgChart Generator
+//
+//  Created by Paul Schmiedmayer on 5/19/20.
+//  Copyright © 2020 Paul Schmiedmayer. All rights reserved.
+//
+
+import SwiftUI
+
+struct ContentView: View {
+    var body: some View {
+        Text("Hello, World!")
+            .frame(maxWidth: .infinity, maxHeight: .infinity)
+    }
+}
+
+
+struct ContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        ContentView()
+    }
+}
