+++ OrgChart/Sources/FaceCrop/Cocoa+Extensions/NSBitmapImageRep.FileType+Extension.swift
