+++ OrgChart/Sources/FaceCrop/CropSquareCenteredOnFaceTransformation.swift
