+++ OrgChart/Tests/FaceCropTests/FaceCropTests.swift
@@ -1,0 +1,8 @@
+import XCTest
+@testable import FaceCrop
+
+final class FaceCropTests: XCTestCase {
+    func testExample() {
+        XCTAssert(true, "✅")
+    }
+}
