+++ OrgChart/Tests/OrgChartTests/OrgChartTests.swift
@@ -1,0 +1,8 @@
+import XCTest
+@testable import OrgChart
+
+final class OrgChartTests: XCTestCase {
+    func testExample() {
+        XCTAssert(true, "✅")
+    }
+}
