--- Tests/LinuxMain.swift
@@ -1,7 +1,0 @@
-import XCTest
-
-import OrgChart_GeneratorTests
-
-var tests = [XCTestCaseEntry]()
-tests += OrgChart_GeneratorTests.allTests()
-XCTMain(tests)
