--- Tests/OrgChartGeneratorTests/OrgChart_GeneratorTests.swift
@@ -1,12 +1,0 @@
-import XCTest
-import class Foundation.Bundle
-
-final class OrgChart_GeneratorTests: XCTestCase {
-    func testExample() throws {
-        print("Running Test")
-    }
-
-    static var allTests = [
-        ("testExample", testExample),
-    ]
-}
